Refactor: Move OutputLogger to common library
In order to be able to log from that library, the logger must be within
it. vscode cannot be imported from the language server so we use
dependency inversion to now pass the output channel to the logger.
